Fix Win64 build
BUG=

Review URL: https://codereview.chromium.org/13463004

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14115 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -851,7 +851,8 @@ RUNTIME_FUNCTION(MaybeObject*, Runtime_ArrayBufferInitialize) {
   Object* byte_length;
   {
     MaybeObject* maybe_byte_length =
-        isolate->heap()->NumberFromDouble(allocated_length);
+        isolate->heap()->NumberFromDouble(
+            static_cast<double>(allocated_length));
     if (!maybe_byte_length->ToObject(&byte_length)) return maybe_byte_length;
   }
   CHECK(byte_length->IsSmi() || byte_length->IsHeapNumber());
